+++ examples/CreateExample.js
@@ -3,18 +3,19 @@
  *
  */
 var basex = require("../index");
+var log = require("../debug");
 //basex.debug_mode = true;
 // create session
 var client = new basex.Session("localhost", 1984, "admin", "admin");
 
 // create new database
-client.create("database", "<x>Hello World!</x>", basex.print);
+client.create("database", "<x>Hello World!</x>", log.print);
 
 // run query on database
-client.execute("xquery /", basex.print);
+client.execute("xquery /", log.print);
 
 // drop database
-client.execute("drop db database", basex.print);
+client.execute("drop db database", log.print);
 
 // close session
 client.close();