+++ examples/QueryBindExample.js
@@ -4,6 +4,7 @@
  */
 
 var basex  = require("../index");
+var log = require("../debug");
 //basex.debug_mode = true;
 var session = new basex.Session("localhost", 1984, "admin", "admin");
 
@@ -12,14 +13,14 @@ var input = "declare variable $name external; for $i in 1 to 10 return element {
 var query = session.query(input);
 
 // bind variable
-query.bind("name", "nodex","",basex.print);
+query.bind("name", "nodex","",log.print);
 
 // print results
-query.execute(basex.print);
+query.execute(log.print);
 // do it again 
-query.bind("name", "again","",basex.print);
-query.execute(basex.print);
-query.info(basex.print);
+query.bind("name", "again","",log.print);
+query.execute(log.print);
+query.info(log.print);
 
 // close query instance
 query.close();